chore: refactor flow ctrl msg and cmd processing
