+++ sn_node/src/node/error.rs
@@ -29,10 +29,16 @@ pub type Result<T, E = Error> = std::result::Result<T, E>;
 pub enum Error {
     #[error("CmdCtrl has stopped.")]
     CmdCtrlStopped,
-    // We shortcircuit and ignore older len prefixes to avoid that
-    // during a split DKG messages are still ongoing post-split
-    // and are sent to the neighbouring section, which causes an AE loop as
-    // section keys are not in chain.
+    /// This should not be possible as the channel is stored in node, and used to process child commands
+    #[error("No more Cmds will be received or processed. CmdChannel senders have been dropped. ")]
+    CmdCtrlChannelDropped,
+    /// This should not be possible as the channel is stored in node, and used to process incoming msgs
+    #[error("No more Msgs will be received or processed. MsgSender has been dropped. ")]
+    MsgChannelDropped,
+    /// We shortcircuit and ignore older len prefixes to avoid that
+    /// during a split DKG messages are still ongoing post-split
+    /// and are sent to the neighbouring section, which causes an AE loop as
+    /// section keys are not in chain.
     #[error("Dkg prefix is shorter than our prefix, so dropping the message.")]
     InvalidDkgPrefix,
     #[error("No membership data exists when it is needed.")]
