+++ sn_node/src/node/flow_ctrl/cmd_ctrl.rs
@@ -18,7 +18,7 @@ use crate::node::{
 
 use custom_debug::Debug;
 use priority_queue::PriorityQueue;
-use std::ops::Add;
+
 use std::time::SystemTime;
 use std::{
     sync::{
@@ -31,7 +31,7 @@ use tokio::{sync::RwLock, time::Instant};
 
 type Priority = i32;
 
-const SLEEP_TIME: Duration = Duration::from_millis(100);
+const EMPTY_QUEUE_SLEEP_TIME: Duration = Duration::from_millis(100);
 
 /// A module for enhanced flow control.
 ///
@@ -72,6 +72,11 @@ impl CmdCtrl {
         self.push_internal(cmd, None).await
     }
 
+    /// Does the cmd_queue contain _anything_
+    pub(crate) fn has_items_queued(&self) -> bool {
+        !self.cmd_queue.is_empty()
+    }
+
     // consume self
     // NB that clones could still exist, however they would be in the disconnected state
     #[allow(unused)]
@@ -86,7 +91,9 @@ impl CmdCtrl {
             return Err(Error::InvalidState);
         }
 
-        let job = CmdJob::new(self.id_counter.add(1), parent_id, cmd, SystemTime::now());
+        self.id_counter += 1;
+
+        let job = CmdJob::new(self.id_counter, parent_id, cmd, SystemTime::now());
 
         let prio = job.priority();
         let _ = self.cmd_queue.push(job, prio);
@@ -104,7 +111,7 @@ impl CmdCtrl {
     }
 
     /// Wait if required by the cmd rate monitoring
-    pub(crate) async fn wait_if_not_processing_at_expected_rate(&mut self) {
+    pub(crate) async fn wait_if_not_processing_at_expected_rate(&self) {
         let expected_rate = self.monitoring.max_cmds_per_s().await;
         let actual_rate = self.attempted.value();
         if actual_rate > expected_rate {
@@ -112,8 +119,8 @@ impl CmdCtrl {
             debug!("Cmd throughput is too high, waiting to reduce throughput");
             log_sleep!(Duration::from_millis((diff * 10_f64) as u64));
         } else if self.cmd_queue.is_empty() {
-            trace!("Empty queue, waiting {SLEEP_TIME:?} to not loop heavily");
-            log_sleep!(SLEEP_TIME);
+            trace!("Empty queue, waiting {EMPTY_QUEUE_SLEEP_TIME:?} to not loop heavily");
+            log_sleep!(EMPTY_QUEUE_SLEEP_TIME);
         }
     }
 
@@ -135,6 +142,7 @@ impl CmdCtrl {
 
         let id = job.id();
         let cmd = job.clone().into_cmd();
+
         let cmd_string = cmd.clone().to_string();
         let priority = job.priority();
